requirements.txt
# In threat-feed-app/requirements.txt
Flask
feedparser
python-dateutil
pandas
openpyxl
